requirements.txt
fastapi==0.111.1
uvicorn==0.30.1
pydantic==2.8.2
python-dotenv==0.20.0
langchain==0.2.12
langchain-google-genai==1.0.8
faiss-cpu==1.8.0
langchain-community==0.2.11
pymupdf==1.24.10
sqlalchemy==2.0.35
psycopg2==2.9.9
